% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_HH_vars_to_opplot_datasets.R
\name{add_HH_vars_to_opplot_datasets}
\alias{add_HH_vars_to_opplot_datasets}
\title{Add HH dataset variables (cluster-related variables) to opplot datasets}
\usage{
add_HH_vars_to_opplot_datasets(VCP = "add_HH_vars_to_opplot_datasets")
}
\arguments{
\item{VCP}{VCQI current program name to be logged, default to be the function name}
}
\value{
Organ pipe plot datasets in VCQI_OUTPUT_FOLDER/Plots_OP augmented with cluster ID and cluster name variables
}
\description{
Add HH dataset variables (cluster-related variables) to opplot datasets
}
