% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_vcqi_databases.R
\name{aggregate_vcqi_databases}
\alias{aggregate_vcqi_databases}
\title{Take all the VCQI databases that can be aggregated and create one single database}
\usage{
aggregate_vcqi_databases()
}
\value{
A dataset consolidating VCQI databases, saved as vcqi_aggregated_databases.rds in VCQI_OUTPUT_FOLDER
}
\description{
Take all the VCQI databases that can be aggregated and create one single database
}
