% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_icc.R
\name{calc_icc}
\alias{calc_icc}
\title{Calculates Intracluster Correlation Coefficient (ICC), confidence interval, and mean squared error for selected outcome}
\usage{
calc_icc(
  y,
  data = NULL,
  svydata = NULL,
  strat = NA,
  cid = NA,
  method = "aov",
  median = FALSE,
  ci.type = "bootstrap",
  ci_bootrep = 500,
  alpha = 0.05,
  showmessages = TRUE
)
}
\arguments{
\item{y}{Outcome variable}

\item{data}{Dataset name (not defined as a survey design object)}

\item{svydata}{Dataset name (defined as a survey design object)}

\item{strat}{Stratum ID variable (when using non-survey design data)}

\item{cid}{Cluster ID variable (when using non-survey design data)}

\item{method}{ICC calculation method}

\item{median}{Use median adjustment to ICC calculation or not, default to be FALSE}

\item{ci.type}{CI calculation method (smith or bootstrap), default to be "bootstrap"}

\item{ci_bootrep}{Number of bootstrap replications if ci.type is bootstrap, default to be 500}

\item{alpha}{Alpha for confidence intervals, default to be 0.05}

\item{showmessages}{Show messages about removed NAs or not, default to be TRUE}
}
\value{
List containing ICC estimate, confidence interval, and mean squared error
}
\description{
Calculates Intracluster Correlation Coefficient (ICC), confidence interval, and mean squared error for selected outcome
}
