% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_interview_date.R
\name{check_interview_date}
\alias{check_interview_date}
\title{Check interview date variable(s) and create VCQI-compatible interview date variables if needed}
\usage{
check_interview_date(dat, errormsgs, exitflag)
}
\arguments{
\item{dat}{Dataset to check for date variables}

\item{errormsgs}{Error messages to pass through (string vector)}

\item{exitflag}{Flag for VCQI error to pass through (a number: 0/1)}
}
\value{
A dataset, saved as <VCQI_RI_DATASET>_preclean in VCQI_OUTPUT_FOLDER; error messages if conditions not met
}
\description{
Check interview date variable(s) and create VCQI-compatible interview date variables if needed
}
