% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/establish_unique_RI_ids.R
\name{establish_unique_RI_ids}
\alias{establish_unique_RI_ids}
\title{Check/create unique ID variables in RI dataset and merge level4 stratifiers}
\usage{
establish_unique_RI_ids(VCP = "establish_unique_RI_ids")
}
\arguments{
\item{VCP}{VCQI current program name to be logged, default to be the function name}
}
\value{
RI_with_ids dataset in VCQI_OUTPUT_FOLDER
}
\description{
Check/create unique ID variables in RI dataset and merge level4 stratifiers
}
\examples{
establish_unique_RI_ids()
}
