% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_svyp_output_database.R
\name{make_svyp_output_database}
\alias{make_svyp_output_database}
\title{Create database with svypd results}
\usage{
make_svyp_output_database(
  variable,
  estlabel,
  vid,
  measureid,
  printprogress = NULL,
  VCP = "make_svyp_output_database",
  ...
)
}
\arguments{
\item{variable}{Outcome variable}

\item{estlabel}{Label for the estimate variable}

\item{vid}{Outcome variable ID for filenames}

\item{measureid}{Analysis indicator ID}

\item{printprogress}{Progress tracking string to print}

\item{VCP}{VCQI current program name to be logged, default to be the function name}

\item{...}{Other arguments}
}
\value{
A database in VCQI_OUTPUT_FOLDER
}
\description{
Create database with svypd results
}
