% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opplot.R
\name{opplot}
\alias{opplot}
\title{Create an organ pipe plot}
\usage{
opplot(
  dat,
  clustvar,
  yvar,
  weightvar = NA,
  stratvar = NA,
  stratum = "",
  barcolor1 = "hotpink",
  barcolor2 = "white",
  linecolor1 = "gray",
  linecolor2 = "gray",
  ylabel = language_string(language_use = language_use, str = "OS_316"),
  ymin = 0,
  ymax = 100,
  yby = 50,
  title = "",
  subtitle = "",
  footnote = "",
  footnotecovg = FALSE,
  footnotedeff = FALSE,
  footnoteicc = FALSE,
  covgcategories = FALSE,
  lowcovgthreshold = NA,
  highcovgthreshold = NA,
  barcolorhigh1 = "#67A9CF",
  barcolormid1 = "#000080",
  barcolorlow1 = "#FF5B00",
  barcolorhigh2 = "#f0f0f0",
  barcolormid2 = "#f0f0f0",
  barcolorlow2 = "#f0f0f0",
  linecolorhigh1 = "gray",
  linecolormid1 = "gray",
  linecolorlow1 = "gray",
  linecolorhigh2 = "gray",
  linecolormid2 = "gray",
  linecolorlow2 = "gray",
  output_to_screen = FALSE,
  filename = NA_character_,
  platform = "png",
  sizew = 7,
  sizeh = 6,
  plotn = FALSE,
  nlinecolor = "black",
  nlinewidth = 0.75,
  nlinepattern = 2,
  ytitle2 = language_string(language_use = language_use, str = "OS_317"),
  yround2 = 5,
  savedata = NA_character_,
  savedatatype = "csv"
)
}
\arguments{
\item{dat}{Dataset to use for plot}

\item{clustvar}{Cluster variable}

\item{yvar}{Outcome variable}

\item{weightvar}{Weight variable}

\item{stratvar}{Strata variable}

\item{stratum}{Specific stratum to plot results for}

\item{barcolor1}{Color for portion of bar where yvar = 1}

\item{barcolor2}{Color for portion of bar where yvar = 0}

\item{linecolor1}{Color for lines separating lower portion of bars}

\item{linecolor2}{Color for lines separating upper portion of bars}

\item{ylabel}{Label for Y axis}

\item{ymin}{Minimum value for Y axis}

\item{ymax}{Maximum value for Y axis}

\item{yby}{Increment for Y axis breaks}

\item{title}{Title of the plot}

\item{subtitle}{Subtitle of the plot}

\item{footnote}{Footnote of the plot}

\item{footnotecovg}{Add estimated coverage as a footnote to the plot}

\item{footnotedeff}{Add design effect (DEFF) as a footnote to the plot}

\item{footnoteicc}{Add intracluster correlation coefficient (ICC) as a footnote to the plot}

\item{covgcategories}{Stratify plot into high, medium, and low coverage categories}

\item{lowcovgthreshold}{The numeric percent value (0-100) used to identify low coverage; clusters with coverage less than or equal to this number will be assigned to the low coverage category}

\item{highcovgthreshold}{The numeric percent value (0-100) used to identify high coverage; clusters with coverage greater than or equal to this number will be assigned to the high coverage category}

\item{barcolorhigh1}{Valid R color or hex color used to represent respondents with yvar = 1 who fall in the high coverage category}

\item{barcolormid1}{Valid R color or hex color used to represent respondents with yvar = 1 who fall in the middle coverage category}

\item{barcolorlow1}{Valid R color or hex color used to represent respondents with yvar = 1 who fall in the low coverage category}

\item{barcolorhigh2}{Valid R color or hex color used to represent respondents with yvar = 0 or missing who fall in the high coverage category}

\item{barcolormid2}{Valid R color or hex color used to represent respondents with yvar = 0 or missing who fall in the middle coverage category}

\item{barcolorlow2}{Valid R color or hex color used to represent respondents with yvar = 0 or missing who fall in the low coverage category}

\item{linecolorhigh1}{Valid R color or hex color used to mark boundaries between the lower portion of the bars in the high coverage category; should have good contrast with barcolorhigh1}

\item{linecolormid1}{Valid R color or hex color used to mark boundaries between the lower portion of the bars in the middle coverage category; should have good contrast with barcolormid1}

\item{linecolorlow1}{Valid R color or hex color used to mark boundaries between the lower portion of the bars in the low coverage category; should have good contrast with barcolorlow1}

\item{linecolorhigh2}{Valid R color or hex color used to mark boundaries between the upper portion of the bars in the high coverage category; should have good contrast with barcolorhigh2}

\item{linecolormid2}{Valid R color or hex color used to mark boundaries between the upper portion of the bars in the middle coverage category; should have good contrast with barcolormid2}

\item{linecolorlow2}{Valid R color or hex color used to mark boundaries between the upper portion of the bars in the low coverage category; should have good contrast with barcolorlow2}

\item{output_to_screen}{Show the plot in plot window or not, default to be FALSE}

\item{filename}{Path to save the plot}

\item{platform}{Type of plot file (may be png, pdf, or wmf; default to png)}

\item{sizew}{Width of the plot file in inches}

\item{sizeh}{Height of the plot file in inches}

\item{plotn}{Plot line showing number of respondents or not, default to be FALSE}

\item{nlinecolor}{Color of line indicating # of respondents}

\item{nlinewidth}{Width of line indicating # of respondents}

\item{nlinepattern}{Pattern of line indicating # of respondents}

\item{ytitle2}{Label for secondary Y axis}

\item{yround2}{Increment for secondary Y axis breaks}

\item{savedata}{Path to save underlying data (if NA, data will not be saved)}

\item{savedatatype}{File type for saving underlying data (may be csv or rds)}
}
\value{
A plot
}
\description{
Create an organ pipe plot
}
