% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RI_CONT_01.R
\name{RI_CONT_01}
\alias{RI_CONT_01}
\title{Dropout between two crude doses}
\usage{
RI_CONT_01(VCP = "RI_CONT_01")
}
\arguments{
\item{VCP}{VCQI current program name to be logged, default to be the function name}
}
\value{
Derived variables, databases, tables, and plots
}
\description{
Dropout between two crude doses
}
\examples{
RI_CONT_01()
}
