% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RI_CONT_01_06PO.R
\name{RI_CONT_01_06PO}
\alias{RI_CONT_01_06PO}
\title{Make plots for RI_CONT_01}
\usage{
RI_CONT_01_06PO(VCP = "RI_CONT_01_06PO")
}
\arguments{
\item{VCP}{VCQI current program name to be logged, default to be the function name}
}
\value{
Plots in VCQI_OUTPUT_FOLDER
}
\description{
Make plots for RI_CONT_01
}
