% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RI_CONT_01B_01PP.R
\name{RI_CONT_01B_01PP}
\alias{RI_CONT_01B_01PP}
\title{Pre-process dataset for RI_CONT_01B}
\usage{
RI_CONT_01B_01PP(VCP = "RI_CONT_01B_01PP")
}
\arguments{
\item{VCP}{VCQI current program name to be logged, default to be the function name}
}
\value{
A dataset (RI_CONT_01B_<ANALYSIS_COUNTER>)
}
\description{
Pre-process dataset for RI_CONT_01B
}
