% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RI_CONT_01B_03DV.R
\name{RI_CONT_01B_03DV}
\alias{RI_CONT_01B_03DV}
\title{Calculate derived variables for RI_CONT_01B}
\usage{
RI_CONT_01B_03DV(VCP = "RI_CONT_01B_03DV")
}
\arguments{
\item{VCP}{VCQI current program name to be logged, default to be the function name}
}
\value{
A dataset (RI_CONT_01B_<ANALYSIS_COUNTER>)
}
\description{
Calculate derived variables for RI_CONT_01B
}
