% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RI_CONT_01B_05TOST.R
\name{RI_CONT_01B_05TOST}
\alias{RI_CONT_01B_05TOST}
\title{Export datasets to Excel for RI_CONT_01B}
\usage{
RI_CONT_01B_05TOST(VCP = "RI_CONT_01B_05TOST")
}
\arguments{
\item{VCP}{VCQI current program name to be logged, default to be the function name}
}
\value{
Sheet(s) in tabular output Excel file in VCQI_OUTPUT_FOLDER
}
\description{
Export datasets to Excel for RI_CONT_01B
}
