% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RI_COVG_01.R
\name{RI_COVG_01}
\alias{RI_COVG_01}
\title{Calculate crude coverage}
\usage{
RI_COVG_01(VCP = "RI_COVG_01")
}
\arguments{
\item{VCP}{VCQI current program name to be logged, default to be the function name}
}
\value{
Derived variables, databases, tables, and plots
}
\description{
Calculate crude coverage
}
\examples{
RI_COVG_01()
}
