% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RI_COVG_02.R
\name{RI_COVG_02}
\alias{RI_COVG_02}
\title{Calculate valid coverage}
\usage{
RI_COVG_02(VCP = "RI_COVG_02")
}
\arguments{
\item{VCP}{VCQI current program name to be logged, default to be the function name}
}
\value{
Derived variables, databases, tables, and plots
}
\description{
Calculate valid coverage
}
\examples{
RI_COVG_02()
}
