% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RI_COVG_03.R
\name{RI_COVG_03}
\alias{RI_COVG_03}
\title{Fully vaccinated}
\usage{
RI_COVG_03(VCP = "RI_COVG_03")
}
\arguments{
\item{VCP}{VCQI current program name to be logged, default to be the function name}
}
\value{
Derived variables, databases, tables, and plots
}
\description{
Fully vaccinated
}
\examples{
RI_COVG_03()
}
