% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RI_COVG_04.R
\name{RI_COVG_04}
\alias{RI_COVG_04}
\title{Not vaccinated}
\usage{
RI_COVG_04(VCP = "RI_COVG_04")
}
\arguments{
\item{VCP}{VCQI current program name to be logged, default to be the function name}
}
\value{
Derived variables, databases, tables, and plots
}
\description{
Not vaccinated
}
\examples{
RI_COVG_04()
}
