% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RI_COVG_04_04GO.R
\name{RI_COVG_04_04GO}
\alias{RI_COVG_04_04GO}
\title{Generate output databases for RI_COVG_04}
\usage{
RI_COVG_04_04GO(VCP = "RI_COVG_04_04GO")
}
\arguments{
\item{VCP}{VCQI current program name to be logged, default to be the function name}
}
\value{
Databases in VCQI_OUTPUT_FOLDER
}
\description{
Generate output databases for RI_COVG_04
}
