% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RI_QUAL_01.R
\name{RI_QUAL_01}
\alias{RI_QUAL_01}
\title{Card and register availability}
\usage{
RI_QUAL_01(VCP = "RI_QUAL_01")
}
\arguments{
\item{VCP}{VCQI current program name to be logged, default to be the function name}
}
\value{
Derived variables, databases, tables, and plots
}
\description{
Card and register availability
}
\examples{
RI_QUAL_01()
}
