% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RI_QUAL_02.R
\name{RI_QUAL_02}
\alias{RI_QUAL_02}
\title{Ever had a card}
\usage{
RI_QUAL_02(VCP = "RI_QUAL_02")
}
\arguments{
\item{VCP}{VCQI current program name to be logged, default to be the function name}
}
\value{
Derived variables, databases, tables, and plots
}
\description{
Ever had a card
}
\examples{
RI_QUAL_02()
}
