% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RI_QUAL_02_02DQ.R
\name{RI_QUAL_02_02DQ}
\alias{RI_QUAL_02_02DQ}
\title{Check data quality for RI_QUAL_02}
\usage{
RI_QUAL_02_02DQ(VCP = "RI_QUAL_02_02DQ")
}
\arguments{
\item{VCP}{VCQI current program name to be logged, default to be the function name}
}
\value{
Errors if conditions not met
}
\description{
Check data quality for RI_QUAL_02
}
