% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RI_QUAL_07B.R
\name{RI_QUAL_07B}
\alias{RI_QUAL_07B}
\title{Calculate valid coverage if there had been no missed opportunities for simultaneous vaccination (MOV)}
\usage{
RI_QUAL_07B(VCP = "RI_QUAL_07B")
}
\arguments{
\item{VCP}{VCQI current program name to be logged, default to be the function name}
}
\value{
Derived variables, databases, tables, and plots
}
\description{
Calculate valid coverage if there had been no missed opportunities for simultaneous vaccination (MOV)
}
\examples{
RI_QUAL_07B()
}
