% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RI_VCTC_01.R
\name{RI_VCTC_01}
\alias{RI_VCTC_01}
\title{RI_VCTC_01: Vaccination Coverage and Timeliness Stacked Bar Charts}
\usage{
RI_VCTC_01(VCP = "RI_VCTC_01")
}
\arguments{
\item{VCP}{VCQI current program name to be logged, default to be the function name}
}
\value{
Datasets, tables, and plots
}
\description{
RI_VCTC_01: Vaccination Coverage and Timeliness Stacked Bar Charts
}
\examples{
RI_VCTC_01()
}
