% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RI_VCTC_01_00GC.R
\name{RI_VCTC_01_00GC}
\alias{RI_VCTC_01_00GC}
\title{Check global macros for RI_VCTC_01}
\usage{
RI_VCTC_01_00GC(VCP = "RI_VCTC_01_00GC")
}
\arguments{
\item{VCP}{VCQI current program name to be logged, default to be the function name}
}
\value{
Errors if conditions not met
}
\description{
Check global macros for RI_VCTC_01
}
