% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RI_VCTC_01_03DV.R
\name{RI_VCTC_01_03DV}
\alias{RI_VCTC_01_03DV}
\title{Calculating bar chart coordinates for VCTC_01}
\usage{
RI_VCTC_01_03DV(VCP = "RI_VCTC_01_03DV")
}
\arguments{
\item{VCP}{VCQI current program name to be logged, default to be the function name}
}
\value{
A dataset with values for a VCTC plot
}
\description{
Calculating bar chart coordinates for VCTC_01
}
