% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RI_VCTC_01_05TO.R
\name{RI_VCTC_01_05TO}
\alias{RI_VCTC_01_05TO}
\title{Export VCTC plot values to Excel}
\usage{
RI_VCTC_01_05TO(VCP = "RI_VCTC_01_05TO")
}
\arguments{
\item{VCP}{VCQI current program name to be logged, default to be the function name}
}
\value{
Sheet(s) in tabular output Excel file in VCQI_OUTPUT_FOLDER
}
\description{
Export VCTC plot values to Excel
}
