% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RI_VCTC_01_06PO.R
\name{RI_VCTC_01_06PO}
\alias{RI_VCTC_01_06PO}
\title{Make plots for RI_VCTC_01}
\usage{
RI_VCTC_01_06PO(VCP = "RI_VCTC_01_06PO_updated")
}
\arguments{
\item{VCP}{VCQI current program name to be logged, default to be the function name}
}
\value{
Plots in VCQI_OUTPUT_FOLDER
}
\description{
Make plots for RI_VCTC_01
}
