% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcqi_check_color.R
\name{vcqi_check_color}
\alias{vcqi_check_color}
\title{Check if a string represents a valid R color}
\usage{
vcqi_check_color(color)
}
\arguments{
\item{color}{String to check}
}
\description{
Check if a string represents a valid R color
}
