% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcqi_multi_lingual_strings.R
\name{vcqi_multi_lingual_strings}
\alias{vcqi_multi_lingual_strings}
\title{Check language choice and language file for MISS VCQI}
\usage{
vcqi_multi_lingual_strings(VCP = "vcqi_multi_lingual_strings")
}
\arguments{
\item{VCP}{VCQI current program name to be logged, default to be the function name}
}
\value{
A global that records the language choice and errors and/or warnings if conditions not met
}
\description{
Check language choice and language file for MISS VCQI
}
\examples{
vcqi_multi_lingual_strings()
}
