% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcqi_to_plot.R
\name{vcqi_to_plot}
\alias{vcqi_to_plot}
\title{Create barplot}
\usage{
vcqi_to_plot(
  database,
  filename,
  datafile = NULL,
  title = NULL,
  name = NULL,
  subtitle = NULL,
  note = NULL,
  caption = NULL,
  savedata = NA,
  savew = 7,
  saveh = 7,
  VCP = "vcqi_to_plot"
)
}
\arguments{
\item{database}{Database used to create the plot}

\item{filename}{Path to save the plot}

\item{datafile}{The file which the database is based on}

\item{title}{Title of the plot}

\item{name}{Plot identifier in file name}

\item{subtitle}{Subtitle of the plot}

\item{note}{Caption of the plot}

\item{caption}{Additional caption (not currently implemented in R)}

\item{savedata}{Path to save data that the plot is based on}

\item{savew}{Width of the plot file in inches}

\item{saveh}{Height of the plot file in inches}

\item{VCP}{VCQI current program name to be logged, default to be the function name}
}
\value{
A plot
}
\description{
Create barplot
}
