% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VCTC_default_global.R
\name{VCTC_default_global}
\alias{VCTC_default_global}
\title{Assign default values for global variables used in RI_VCTC_01}
\usage{
VCTC_default_global()
}
\value{
Values in the global environment
}
\description{
Assign default values for global variables used in RI_VCTC_01
}
\examples{
VCTC_default_global()
}
