% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals_for_timeliness_plots -
%   modified_legend_for_bcg.R
\name{VCTC_global_modified_legend_for_bcg}
\alias{VCTC_global_modified_legend_for_bcg}
\title{Assign default and some customized values for global variables used in RI_VCTC_01}
\usage{
VCTC_global_modified_legend_for_bcg()
}
\value{
Values in the global environment
}
\description{
Assign default and some customized values for global variables used in RI_VCTC_01
}
\examples{
VCTC_global_modified_legend_for_bcg()
}
