% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals_for_timeliness_plots -
%   same_legend_for_all_doses.R
\name{VCTC_global_same_legend_for_all}
\alias{VCTC_global_same_legend_for_all}
\title{Assign default values for global variables used in RI_VCTC_01}
\usage{
VCTC_global_same_legend_for_all()
}
\value{
Values in the global environment
}
\description{
Assign default values for global variables used in RI_VCTC_01
}
\examples{
VCTC_global_same_legend_for_all()
}
