% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_RI_augmented_dataset.R
\name{make_RI_augmented_dataset}
\alias{make_RI_augmented_dataset}
\title{Merge RI indicators calculated by VCQI onto the originally provided RI dataset}
\usage{
make_RI_augmented_dataset(outpath = NA)
}
\arguments{
\item{outpath}{Path where RI indicator datasets are saved (by default the function looks in VCQI_OUTPUT_FOLDER)}
}
\value{
RI_augmented_dataset and vcqi_ads_logfile datasets in VCQI_OUTPUT_FOLDER
}
\description{
Merge RI indicators calculated by VCQI onto the originally provided RI dataset
}
\examples{
make_RI_augmented_dataset(outpath = NA)
}
