% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opplot.R
\name{opplot}
\alias{opplot}
\title{Create an organ pipe plot}
\usage{
opplot(
  dat,
  clustvar,
  yvar,
  weightvar = NA,
  stratvar = NA,
  stratum = "",
  barcolor1 = "hot pink",
  barcolor2 = "white",
  linecolor1 = "gray",
  linecolor2 = "gray",
  ylabel = language_string(language_use = language_use, str = "OS_316"),
  ymin = 0,
  ymax = 100,
  yby = 50,
  title = "",
  subtitle = "",
  footnote = "",
  output_to_screen = FALSE,
  filename = NA_character_,
  platform = "png",
  sizew = 7,
  sizeh = 6,
  plotn = FALSE,
  nlinecolor = "black",
  nlinewidth = 0.75,
  nlinepattern = 2,
  ytitle2 = language_string(language_use = language_use, str = "OS_317"),
  yround2 = 5,
  savedata = NA_character_,
  savedatatype = "csv"
)
}
\arguments{
\item{dat}{Dataset to use for plot}

\item{clustvar}{Cluster variable}

\item{yvar}{Outcome variable}

\item{weightvar}{Weight variable}

\item{stratvar}{Strata variable}

\item{stratum}{Specific stratum to plot results for}

\item{barcolor1}{Color for portion of bar where yvar = 1}

\item{barcolor2}{Color for portion of bar where yvar = 0}

\item{linecolor1}{Color for lines separating lower portion of bars}

\item{linecolor2}{Color for lines separating upper portion of bars}

\item{ylabel}{Label for Y axis}

\item{ymin}{Minimum value for Y axis}

\item{ymax}{Maximum value for Y axis}

\item{yby}{Increment for Y axis breaks}

\item{title}{Title of the plot}

\item{subtitle}{Subtitle of the plot}

\item{footnote}{Footnote of the plot}

\item{output_to_screen}{Show the plot in plot window or not, default to be FALSE}

\item{filename}{Path to save the plot}

\item{platform}{Type of plot file (may be png, pdf, or wmf; default to png)}

\item{sizew}{Width of the plot file in inches}

\item{sizeh}{Height of the plot file in inches}

\item{plotn}{Plot line showing number of respondents or not, default to be FALSE}

\item{nlinecolor}{Color of line indicating # of respondents}

\item{nlinewidth}{Width of line indicating # of respondents}

\item{nlinepattern}{Pattern of line indicating # of respondents}

\item{ytitle2}{Label for secondary Y axis}

\item{yround2}{Increment for secondary Y axis breaks}

\item{savedata}{Path to save underlying data (if NA, data will not be saved)}

\item{savedatatype}{File type for saving underlying data (may be csv or rds)}
}
\value{
A plot
}
\description{
Create an organ pipe plot
}
